+++ scripts/module.js
@@ -17,28 +17,29 @@ import { setupSocket } from "./GMAction.js";
 import { setupBetterScrollCreation } from "./betterScrollCreation.js";
 import { setupGriddedGridless } from "./griddedGridless.js";
 import { initActorDispositionColors } from "./ActorDispositionColors.js";
+import { setupContainerHelpers } from "./containerHelper.js";
 
 export let systemString;
 export let systemConfig;
 export let localizeHeader;
+export let libWrapper;
 
 Hooks.once("init", async function () {
+  libWrapper = globalThis.libWrapper;
 	console.log("dnd5e-scriptlets | doing init setup");
 	setProperty(globalThis, "dnd5eScriptlets.api", {});
   registerSettings();
 	setupTokenResizer();
   systemString = game.system.id;
+  systemConfig = game.system.config;
   switch (systemString) {
     case "dnd5e":
-      systemConfig = CONFIG.DND5E;
       localizeHeader = "DND5E";
       break;
     case "sw5e":
-      systemConfig = CONFIG.SW5E;
       localizeHeader = "SW5E";
       break;
     case "dcc": 
-    systemConfig = CONFIG.DCC;
     localizeHeader = "DCC";
     break;
   }
@@ -72,6 +73,7 @@ Hooks.once("ready", async function () {
   setupUpdateCreatedOrigins();
   setupBetterScrollCreation();
   setupGriddedGridless();
+  setupContainerHelpers();
   const module = game.modules.get("dnd5e-scriptlets");
   if (module) {
     module.api = globalThis.dnd5eScriptlets.api;